requirements.txt
fastapi
uvicorn
python-dotenv
requests
razorpay
twilio
pydantic
supabase